Power Manager refactor
Big commit. Probably too big. Added a power manager class. Also hooked
everything up using the clean sweep class as a sort of controller for
the whole project. Also fixed sample XML file.

+++ src/main/java/PowerManagement/IPowerManager.java
@@ -2,9 +2,9 @@ package main.java.PowerManagement;
 
 public interface IPowerManager {
 
-	public int GetBatteryLevel();
+	public double GetBatteryLevel();
 	
-	public boolean RequestEnergy(int requestedEnergyAmount);
+	public boolean RequestEnergy(double requestedEnergyAmount);
 	
 	//once the above is complete, we should find more graceful ways of determining
 	//home much energy is needed
